+++ src/components/CustomTable/StyledTableCell/StyledTableCell.js
@@ -1,6 +1,6 @@
 import { TableCell, withStyles } from "@material-ui/core"
 
-import replaceIfNull from "@/utils/replaceIfNull"
+import replaceIfNull from "@/utils/replaceIfNullUtils"
 
 const NewTableCell = withStyles((theme) => ({
   root: {
